Refactor: nuke two {-# SOURCE #-} imports of TC.Monad.Base
